Update Main.kt
Updated delteNote() function

This function now displays all the notes by calling the listNotes() function.
Then we ask the user to enter an index of the note to delete and then we validate if the delete was successful or not to display a message in the console for the user.