new functions
Added two new functions.

First function was added into Main.kt.
This function asks the user for the index of the note to archive.

Second function was added into NoteAPI.kt
This function archiveNote() is a boolean function, it returns either true or false. It takes in an index and verifies if the the note of that index exists. If the note exists then change the isNotArchived boolean to true and returns true, else return false.