Update Main.kt
Added new function to search by title.

This function prompts the user to enter a title of the note the user wants to search for. Then this title is being passed in to a searchForNote function in the NoteAPI.kt, which will return a list or the note that contains the title.