new utility for category validation
This utility will allow us to validate the category of the note.

The CategoryUtility.kt has a set of categories from which the user needs to pick from, then we have a function which uses that set to validate if the note category is equal to one of the categories in the set.